[REFACTOR/#9] 리뷰 기반 리팩토링
